+++ packages/wix-storybook-utils/src/AnchoredTitle/styles.scss
@@ -1,5 +1,5 @@
-@import '../../ui/typography.scss';
-@import '../../ui/colors.scss';
+@import '../ui/typography.scss';
+@import '../ui/colors.scss';
 
 .titleContainer {
   cursor: pointer;
